5-flip_bits.c fxd ul

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -12,8 +12,8 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 
 {
-	unsigned int comp = 0;
-	unsigned int cont = 0;
+	unsigned long int comp = 0;
+	int cont = 0;
 
 	comp = n ^ m;
